Added ACM Discord Link to Stay Connected Section and Refactored MailingListBtn and MailingListBanner
