--- src/components/JoinMailingListBtn/JoinMailingListBtn.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Button from '@material-ui/core/Button';
-
-function JoinMailingListBtn() {
-	return <Button
-		color="secondary"
-		variant="contained"
-		href="http://eepurl.com/c5pE6P"
-		target="_blank"
-		rel="noopener noreferrer"
-	>
-		Join Our Mailing List
-	</Button>;
-}
-
-export default JoinMailingListBtn;
